A big rewrite on how we do class setting that tries to set types based on expression types.
Let tree nodes be able to return the type of the expression.
Change how ClassSetter deals with things so that it first scans
fields and methods in classes and sets up Scopes with variables.
Then it tris to evaluate types on expression, which may need the
fields and variables we have in the Scopes.

FieldAndMethodScanner became merged into ClassSetter so it is no more.