Change type setting for anonymous classes so that it works even for typed classes.
Inner classes created from a field or another type need to know the
outer class, so we have to set it. This meant that we could not send
anonymousClass-messages without the from-clause. For easier handling
of the events UntypedClassInstanceCreationExpression is not used in
the tree, we now only build ClassInstanceCreationExpression in the final
tree and this means that we always have access to the from-clause.
This also sets type for cast expressions.