Implement bytecode generation for if.
Stopped trying to replace DottedName on a case by case basis and
instead let DottedName have an inner tree that we set in ClassSetter,
this means we can replace in one place only. This also means that we
have to update DottedName with visit and getChildNodes methods.

Implement visitor for TwoPartExpression and UnaryExpression so that we
can do comparisson in the ByteCodeGenerator.