Try to handle statically imported fields.
In order to make this work I changed Identifier so that it can hold an
actual treenode since Identifier becomes FieldAccess for some fields.
If the actual node is set then Identifier will delegate tree walking
calls to it. This change also meant that I could remove the
setFrom/setOn methods.

Also a start of dealing with lambda expressions.